Depend on Username
With some craftbukkit changes I made, I added a username field that is
immutable which is set to the player's correct name when the
EntityPlayer is created. This is so that I can freely modify the players
names without having to store a clean copy else where. Finally, in the
PlayerQuitEvent, I reset the name back to its correct value so the
player can properly save.