LOTS OF FIXES
Fixed player death being called a LOT when they die.
Fixed jump particles from showing up.
Fixed Aspect data from OVERWRITING itself when it shouldn't.
Reorganized health so that updateHealth really only just updates it.
^ That also means that damage and heal will call the particle effects
now.
The unmoveable aspect key now does something.
Added a caching of the jump/speed modifiers, this means it only needs to
update it on the player when it really changes and not CONSTANTLY.