More bug fixes
This time around, I added some safeguards for setting the prefix and
suffix so it doesn't spill over 16 characters for either.
After that is the moving of the prefix setter to onJoin from onCreation.
The player's connection is not yet there is onCreation so therefore we
need to use onJoin.