Call/Improve ButtonClickEvent and buttons
First thing's first, having the button flag on an item will remove the
unmoveable flag on the item. This is because the button flag will result
in the item being unmoveable. I don't want players to see both the
"button" and "unmoveable" flags in the lore. I made buttons unmoveable
because of the nature of buttons. Why would you even be able to move a
button for pete's sake. It's not right.