Improve Stockpile check speed, improve DragEvent handling
Replaced the string check to an instanceof for checking if an
inventory's holder is a stockpile. Tests show instanceof is 2.5 times
faster on the low end.

For improving the drag event handling, I added a drag event to
stockpiles. When a drag event occurs in the stockpile inventory, the
onDrag method is called.